Review tidy-up: break circular import, use FP for group/merge
